# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handlers.CreateDepartmentRequest:
    properties:
      contact_email:
        example: waste@islamabad.gov
        type: string
      description:
        example: Handles all waste-related issues
        type: string
      name:
        example: Waste Management
        type: string
    required:
    - name
    type: object
  handlers.CreateNotificationRequest:
    properties:
      message:
        example: Your report is now in review
        type: string
      report_id:
        type: string
      title:
        example: Status Update
        type: string
      type:
        enum:
        - status_update
        - safety_alert
        - nearby_issue
        example: status_update
        type: string
      user_id:
        type: string
    required:
    - message
    - title
    - type
    - user_id
    type: object
...
  handlers.UpdateDepartmentRequest:
    properties:
      contact_email:
        type: string
      description:
        type: string
      name:
        example: Waste Management
        type: string
    type: object
...
  models.Department:
    properties:
      contact_email:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
...
  models.Notification:
    properties:
      created_at:
        type: string
      id:
        type: string
      is_read:
        type: boolean
      message:
        type: string
      report_id:
        type: string
      title:
        type: string
      type:
        $ref: '#/definitions/models.NotificationType'
      user_id:
        type: string
    type: object
  models.NotificationType:
    enum:
    - status_update
    - safety_alert
    - nearby_issue
    type: string
    x-enum-varnames:
    - NotifStatusUpdate
    - NotifSafetyAlert
    - NotifNearbyIssue
...
  utils.APIResponse:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
  utils.PaginatedResponse:
    properties:
      data: {}
      page:
        type: integer
      page_size:
        type: integer
      success:
        type: boolean
      total_items:
        type: integer
      total_pages:
        type: integer
    type: object
host: localhost:8080
info:
  contact:
    email: zaid@example.com
    name: Zaid Ejaz
  description: Backend API for the Saaf Islamabad civic issue reporting platform.
    Citizens can report issues like garbage, broken streetlights, and road damage.
    Admins assign issues to department staff who resolve them.
  title: Saaf Islamabad API
  version: "1.0"
paths:
...
  /departments:
    get:
      description: Get all departments
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Department'
                  type: array
              type: object
      summary: List departments
      tags:
      - Departments
    post:
      consumes:
      - application/json
      description: Add a new department (admin only)
      parameters:
      - description: Department data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateDepartmentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Department'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Create department
      tags:
      - Departments
  /departments/{id}:
    delete:
      description: Remove a department (admin only)
      parameters:
      - description: Department UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Delete department
      tags:
      - Departments
    get:
      description: Get department by ID
      parameters:
      - description: Department UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Department'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.APIResponse'
      summary: Get department
      tags:
      - Departments
    put:
      consumes:
      - application/json
      description: Update department details (admin only)
      parameters:
      - description: Department UUID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateDepartmentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Department'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Update department
      tags:
      - Departments
  /notifications:
    get:
      description: Get notifications for the authenticated user
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: page_size
        type: integer
      - description: Only unread
        in: query
        name: unread
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.PaginatedResponse'
      security:
      - BearerAuth: []
      summary: List my notifications
      tags:
      - Notifications
    post:
      consumes:
      - application/json
      description: Send a notification to a user (admin/system)
      parameters:
      - description: Notification data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateNotificationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Notification'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Create notification
      tags:
      - Notifications
  /notifications/{id}/read:
    patch:
      description: Mark a single notification as read
      parameters:
      - description: Notification UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Mark notification as read
      tags:
      - Notifications
  /notifications/read-all:
    patch:
      description: Mark all notifications as read for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Mark all notifications as read
      tags:
      - Notifications
...
securityDefinitions:
  BearerAuth:
    description: 'Type "Bearer" followed by a space and your JWT token. Example: **Bearer
      eyJhbGci...**'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"